PLC4X-257 Swap anonymous lambdas which static classes.
Lambdas which are present on stack trace have a big impact on debugging capabilities.
They do not point place in code which is faulty, instead they represent kind of named state for code which formed it.
In order to improve overall developer experience swap to regular classes gives us a possiblity to find which `only` call refused message from further processing.