Kafka connect
- Minor changes to source connector allow for empty config. This allows 
to add connectors to Kafka that don't have any source or jobs config.

- Add Confluent maven plugin to package connector for Confluent Hub.

- Changed example configs to use the simulated driver instead of S7. 
This allows for a defaut connector to be added that works without 
external PLCs.

- Removed the ConfigDef define calls, this allows for an empty connector 
to be created. I am also having trouble with the Config.Type.LIST types 
resultig in a NullPointerException when creating connectors in 
Confluent.