First Pass at allowing expressions for string types in mspec.
This allows you to assign an expression for the size of a string. It is 
useful for pascal strings where the size precedes the string.

I'm not happy with not using the sizeInBits field in 
DefaultStringTypeReference. It may be better to change the sizeInbits 
field to a String type for all data types instead of just string?