PLC4X-323 Second attempt to fix memory issues with socketcan transport.
Given that buffer is re-used earlier usage with reset() call was inacurate.
It did change buffer state and blocking further use.
Fix which rely on rewind moves writer index to beginning of buffer making new data override old one.
Also as it is only position change it does not involve re-allocation of memory.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>