plc4j/spi: Fix readString() on non-ascii encodings
The previous logic could result in a `StringIndexOutOfBoundsException`
upon calling `substring()`.
Reason is that the calculated `realLength` is in bytes and the length
of the created string is the number of characters. For UTF encodings,
multiple bytes can be just one character. This causes `realLength`
to be longer than the actual string and thus the `substring()`-call
to fail.

Fix this by applying `realLength` to the byte-array instead of the
string. This is done by utilizing a different String constructor.

Note that on the UTF16 cases the `realLength` needs to be incremented
by 2 to account for the two added bytes.

This commit contains a regression test that fails without these
adaptations.

Signed-off-by: Marc Aurel Fritz <Marc-Aurel.Fritz@plansee.com>