Fix reading UTF-8 strings (from OPC UA nodes) (#832)
* plc4j/spi: Fix readString() on non-ascii encodings

The previous logic could result in a `StringIndexOutOfBoundsException`
upon calling `substring()`.
Reason is that the calculated `realLength` is in bytes and the length
of the created string is the number of characters. For UTF encodings,
multiple bytes can be just one character. This causes `realLength`
to be longer than the actual string and thus the `substring()`-call
to fail.

Fix this by applying `realLength` to the byte-array instead of the
string. This is done by utilizing a different String constructor.

Note that on the UTF16 cases the `realLength` needs to be incremented
by 2 to account for the two added bytes.

This commit contains a regression test that fails without these
adaptations.

Signed-off-by: Marc Aurel Fritz <Marc-Aurel.Fritz@plansee.com>

* plc4j/spi: Always convert encoding to upper case at readString()

Previously, the inner switch statement at the UTF-16 cases
would not convert encoding to upper case. This prevented
matching from working properly.
Do so directly at the beginning in order to prevent future
bugs.

Signed-off-by: Marc Aurel Fritz <Marc-Aurel.Fritz@plansee.com>

* plc4j/spi: Merge ASCII and UTF-8 cases on readString()

Prevent having the same logic twice.

Signed-off-by: Marc Aurel Fritz <Marc-Aurel.Fritz@plansee.com>

---------

Signed-off-by: Marc Aurel Fritz <Marc-Aurel.Fritz@plansee.com>