test(plc4go/cbus): fix concurrency issue in test.
Transition away from using WaitGroup on addResponseCode or addPlcValue and hook a chan close into transaction end with EndTransaction or FailTransaction as this is really the marker which notes if the transaction is ended (what we looking for at the end). This has two benefits: 1. No more flaky tests and 2. if we forget to properly end a transaction we will notice and need to fix that in prod code.