fix(plc4j): add fireDiscoverEvent (replacing awaitDiscoverComplete) and moving awaitDiscoverComplete to it's true prupose
Rational: logic is wrong here. The awaits primary purpose it for the Driver test runner to not block. So the action should be done in any case just the waiting should be controlled. As you can see here the event in only fired if the variable is through hence affecting logic. So there should be a new variable called fire discover and the keep the current one to wait for something