fix(plc4j): Adjust encoding/decoding logic to support mixed key lengths.
As found during the tests of MX OPC server, our encoding logic did not work properly when client and server used different key lengths.
Changes introduced in this commit address this issue, and add few additional tests to confirm valid behavior.

Relates to #1682 and #1802, whereas second issue lead to discovery of the bug.

Signed-off-by: Łukasz Dywicki <luke@code-house.org>